# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Book-RESTful-API",
        "description": "Login to obtain a JWT, with bearer authorization",
        "contact": {
            "email": "demo@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/books": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "取得最多25筆的書本與評分",
                "operationId": "a9d13283dfab262e97af3c914d51b7a8",
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Book"
                ],
                "summary": "創建書本",
                "operationId": "47047e95a700b3271b697e57d0965191",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "標題",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "簡介",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "創建成功"
                    },
                    "401": {
                        "description": "未驗證"
                    },
                    "404": {
                        "description": "資源不存在"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/books/{bookId}": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "取得書本",
                "operationId": "5c80e8a28c94561863abe0e07bc89857",
                "parameters": [
                    {
                        "name": "bookId",
                        "in": "path",
                        "description": "Book id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "404": {
                        "description": "資源不存在"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Book"
                ],
                "summary": "刪除書本",
                "operationId": "101f4fe941f5452bd4aed15444311714",
                "parameters": [
                    {
                        "name": "bookId",
                        "in": "path",
                        "description": "Book id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "刪除成功",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "資源不存在",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Book"
                ],
                "summary": "更新書本",
                "description": "更新書本",
                "operationId": "1fa908e346ccb89d1c05d1735a06bdf3",
                "parameters": [
                    {
                        "name": "bookId",
                        "in": "path",
                        "description": "Book id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "更新書本的內容",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "標題",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "內容",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "未驗證",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "資源不存在",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Bearer jwt token",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}